\name{Assignment 3-package}
\alias{Assignment 3-package}
\alias{Assignment 3}
\docType{package}
\title{
\packageTitle{Assignment 3}
}
\description{
\packageDescription{Assignment 3}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{Assignment 3}
\packageIndices{Assignment 3}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{Assignment 3}

Maintainer: \packageMaintainer{Assignment 3}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
